:art: Flyttet toc-logikk ut i hook
